fix: 主题切换元服务改成元数据形式
